Set up protorype
I set up a prototype classifier that will load the data, preprocess it, perform feature selection, and then run a random forest model for classification.
I separated the loading an preprocessing features into a separate file - data_loader.py. I left the the functionality in main for now as I was not sure of the best way to split it up.